third_party/WebKit/LayoutTests/fast/forms/select/menulist-popup-type-ahead-style-change-expected.txt

PASS internals.selectMenuListText(select) is "C"
PASS successfullyParsed is true

TEST COMPLETE
Tests <select> style change when popup is visible after selection is changed by keyboard typeahead. The selection should not change when style changes. Needs run-layout-test.
